+++ internal/model/dict_item.go
@@ -1,7 +1,7 @@
 package model
 
 type DictItem struct {
-	ID       int    `gorm:"primaryKey" json:"id"`
+	ID       uint64 `gorm:"primaryKey" json:"id"`
 	DictCode string `gorm:"index;not null" json:"dictCode"`            // 外键，关联 Dict
 	Value    string `gorm:"type:varchar(200);not null" json:"value"`   // 数据值，系统用的
 	LabelZh  string `gorm:"type:varchar(200);not null" json:"labelZh"` // 中文名称
