+++ internal/model/role.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Role struct {
-	ID          int           `gorm:"primaryKey;autoIncrement" json:"id"`
+	ID          uint64        `gorm:"primaryKey;autoIncrement" json:"id"`
 	Name        string        `gorm:"type:varchar(30);not null" json:"name"`
 	Users       []*User       `gorm:"many2many:user_roles" json:"users"`
 	Permissions []*Permission `gorm:"many2many:role_permissions" json:"permissions"`
